# lcm-engine/lcm_engine/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: |
    The Life-Cycle Management Engine API specification for the Gaia-X project.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: LCM Engine API
  version: 2.3.0
servers:
- description: Testing LCM Engine server 1
  url: https://demo.gxfs.xlab.si
- description: Testing LCM Engine server 2
  url: https://orchestrator.gxfs.xlab.si
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /secret:
    get:
      operationId: get_secrets
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Secret'
                type: array
          description: A list of secrets
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
      summary: List available secrets
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
    post:
      operationId: create_secret
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/SecretBody'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
          description: The created secret
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
      summary: Create a new secret
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
  /secret/{secretId}:
    delete:
      operationId: delete_secret
      parameters:
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Secret deleted
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Delete a secret
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
    get:
      operationId: describe_secret
      parameters:
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
          description: A secret
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Describe a secret
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
    patch:
      operationId: update_secret
      parameters:
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/SecretBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
          description: The updated secret
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Update a secret
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
    put:
      operationId: replace_secret
      parameters:
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/SecretBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
          description: The updated secret
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Replace a secret
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
...
  /workspace/{workspaceId}/secret:
    get:
      operationId: list_workspace_secrets
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Secret'
                type: array
          description: A list of secrets
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: List secrets assigned to the workspace
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
  /workspace/{workspaceId}/secret/{secretId}:
    delete:
      operationId: remove_workspace_secret
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Secret deleted
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Remove the secret from the workspace
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
    put:
      operationId: assign_secret
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: secretId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Secret assigned to workspace
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Assign a new secret to the workspace
      tags:
      - secret
      x-openapi-router-controller: lcm_engine.controllers.secret_controller
components:
...
  requestBodies:
    SecretBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Secret'
      description: Secret specification object
      required: true
...
  schemas:
...
    Secret:
      additionalProperties: false
      description: A secret describing environment variables and/or files with confidential
        information
      example:
        file:
          path: /root/.config/openstack.yaml
          contentsHash: |
            9b71d224bd62f3785d96d46ad3ea3d73
            319bfbc2890caadae2dff72519673ca7
            2323c3d99ba5c11d7c7acc6e14b8c5da
            0c4663475c2e5c3adef46f73bcdec043
          contents: contents
        name: my secret
        id: 6
        workspaces:
        - 3
        - 6
        - 8
        env:
          MY_ENV_ID: 123abc
          MY_ENV_KEY: h4rd2br34km3
      properties:
        id:
          example: 6
          format: int64
          readOnly: true
          title: id
          type: integer
        name:
          example: my secret
          title: name
          type: string
        workspaces:
          example:
          - 3
          - 6
          - 8
          format: int64
          items:
            type: integer
          readOnly: true
          title: workspaces
          type: array
        file:
          $ref: '#/components/schemas/Secret_file'
        env:
          additionalProperties:
            minProperties: 1
            nullable: true
            type: string
          example:
            MY_ENV_ID: 123abc
            MY_ENV_KEY: h4rd2br34km3
          title: env
          type: object
      required:
      - id
      - name
      - workspaces
      title: Secret
      type: object
...
    Error:
      additionalProperties: false
      description: An error
      properties:
        msg:
          title: msg
          type: string
      required:
      - msg
      title: Error
      type: object
...
    Secret_file:
      additionalProperties: false
      example:
        path: /root/.config/openstack.yaml
        contentsHash: |
          9b71d224bd62f3785d96d46ad3ea3d73
          319bfbc2890caadae2dff72519673ca7
          2323c3d99ba5c11d7c7acc6e14b8c5da
          0c4663475c2e5c3adef46f73bcdec043
        contents: contents
      properties:
        path:
          example: /root/.config/openstack.yaml
          title: path
          type: string
        contents:
          description: Base64-encoded contents
          format: byte
          title: contents
          type: string
          writeOnly: true
        contentsHash:
          description: The SHA3-512 checksum of the contents encoded as a hexadecimal
            string
          example: |
            9b71d224bd62f3785d96d46ad3ea3d73
            319bfbc2890caadae2dff72519673ca7
            2323c3d99ba5c11d7c7acc6e14b8c5da
            0c4663475c2e5c3adef46f73bcdec043
          readOnly: true
          title: contentsHash
          type: string
      title: Secret_file
      type: object